requirements.txt
# Core dependencies
streamlit>=1.29.0
pandas>=2.0.0
numpy>=1.24.0

# Vector database and embeddings
pinecone>=5.0.0
sentence-transformers>=2.5.0
huggingface-hub>=0.20.0

# LLM integration
google-generativeai>=0.3.0

# Utilities
python-dotenv>=1.0.0
tiktoken>=0.5.0

# Optional: For better performance
torch>=2.0.0
transformers>=4.36.0
